indentation; no gui reloading; timer for main loop instead of a loop
Indentation changed to 4 spaces.

Gui destroying is mitigated by introducing a proxy to `Gui, Add` which adds
a control OR if the control already exists, just moves it. All existing
controls are hidden before the level starts.

!Note: Small artifact: on level two the blocks seems to have borders. This
is probaly due to the controls being hidden and re-shown. Also the last
enemy appears white (no idea why).

Whenever a script is being resized/moved or when a menu is shown the
current thread is interrupted until whatever it was is completly finished.
Since a delta is being used to time the frames this caused that single frame
to have a huge delta which rocketed the player out of bounds due to gravity.

Using a timer circumvents this issue.